Added a 'fmradio' example using the FMCOMMS2 source block
For it to work properly, you may have to switch the "gain control mode"
to "fast attack" with the IIO oscilloscope, then to "slow attack" again.

It is advised to run the generated Python script with a negative
niceness and/or with a realtime kernel, to ensure that no samples are
lost.

Signed-off-by: Paul Cercueil <paul.cercueil@analog.com>