Add a GNU Radio based fmcomms_transceiver application
This application requires the following:
* A FMCOMMS 2, 3 or 4,
* GNU Radio,
* the 'gr-baz' blocks for GNU Radio,
* the 'gr-iio' blocks for GNU Radio,
* super-user rights (for creating the TAP interface).

Once launched, you will have a new 'tap0' network interface, which you
can use to communicate with another board equipped with a FMCOMMS card.

The program is generated by the GNU Radio Companion Compiler (grcc).
The flow-graphs with the *.grc extension correspond to the sources, and
can be viewed and/or edited using the GNU Radio Companion utility.

Signed-off-by: Paul Cercueil <paul.cercueil@analog.com>