Math block: Use weak pointers in block list
Previously, some of the blocks generated had a shared pointer to the
iio_math block. What happened, is that those shared pointers prevented
the main iio_math object to be destroyed, resulting in a leak of memory.
To avoid that, we now use weak pointers.

Besides, the "block" C structure was allocated in various places but
never freed. They are now registered in the iio_math object for later
deletion.

Signed-off-by: Paul Cercueil <paul.cercueil@analog.com>