Math block: Do NOT use weak pointers in block list
The problem of using weak pointers is that there must be an instance of
a shared_ptr for the reference to stay valid, which we cant ensure here.

What we do now, is free the block list at the end of the constructor,
after all the blocks have been connected to the flowgraph.

Signed-off-by: Paul Cercueil <paul.cercueil@analog.com>